# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.2",
    "title": "WAFRN Backend",
    "description": "API routes for wafrn social network"
  },
  "servers": [
    {
      "url": "http://0.0.0.0:3000/"
    }
  ],
  "paths": {
...
    "/api/block": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/unblock": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/follow": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/unfollow": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/uploadMedia": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "adultContent": {
                    "example": "any"
                  },
                  "nsfw": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/updateMedia": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "adultContent",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NSFW",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/singlePost/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/blog": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/createPost": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "captchaKey": {
                    "example": "any"
                  },
                  "parent": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  },
                  "content_warning": {
                    "example": "any"
                  },
                  "privacy": {
                    "example": "any"
                  },
                  "tags": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reportPost": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "severity": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "postId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}